# web/openapi/api.yml
openapi: 3.1.0
info:
  title: API spec, mainly for client generation
  version: 0.0.1
paths:
...
  "/api/items/reset-to-preset":
    post:
      summary: Reset items to preset
      operationId: resetToPreset
      tags:
        - Items
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                 type: string
              required:
                - name
      responses:
        "200":
          description: OK
...